[#32] refactor: 모달 비페이지화, 컴포넌트 그룹핑
