--- app/(participateModal)/@modal/(.)participate/page.tsx
@@ -1,8 +1,0 @@
-import { NextPage } from "next";
-import ParticipatePetGroupModal from "@/components/ParticipatePetGroupModal/ParticipatePetGroupModal";
-
-const ParticipatePetGroupPage: NextPage = () => {
-  return <ParticipatePetGroupModal />;
-};
-
-export default ParticipatePetGroupPage;
