--- app/(participateModal)/layout.tsx
@@ -1,10 +1,0 @@
-import { ReactNode } from "react";
-
-export default function SelectLayout({ children, modal }: { children: ReactNode; modal: ReactNode }) {
-  return (
-    <>
-      {children}
-      {modal}
-    </>
-  );
-}
