+++ components/Home/HomeHealthLogPreview/HomeHeathLogPreview.tsx
@@ -1,10 +1,10 @@
-import HealthLogTaskList from "../@common/LogList";
+import LogList from "@/components/@common/LogList";
 
 const HomeHeathLogPreview = () => {
   return (
     <>
       <p>⭐️ 나나의 스케줄을 잊지 마세요!</p>
-      <HealthLogTaskList pageType="home" />
+      <LogList pageType="home" />
     </>
   );
 };
